fix(spec): IntegerArgumentType.valid_range is tuple[int, int]; drop runtime int() truncation

+++ mav_gss_lib/platform/spec/argument_types.py
@@ -32,7 +32,7 @@ class IntegerArgumentType:
     signed: bool = False
     byte_order: ByteOrder = "little"
     description: str = ""
-    valid_range: tuple[float, float] | None = None
+    valid_range: tuple[int, int] | None = None
     valid_values: tuple[int, ...] | None = None
